refactoring and making mobile friendly
