--- src/client/assets/style/_airport-picker-block.scss
@@ -1,13 +1,0 @@
-.airport-picker {
-  font-weight: 600;
-  width: 40%;
-  //height: 40px;
-  margin: 0px 10px;
-}
-
-@media screen and (max-width: 479px) {
-  .airport-picker {
-    width: 100%;
-    margin: 0;
-  }
-}
