--- src/client/assets/style/_ticket-details-block.scss
@@ -1,36 +1,0 @@
-.ticket-details-block-modal-content {
-  //border: 1px solid black;
-  border-radius: 5px;
-  position: relative;
-  top: 10em;
-  background-color: #fff;
-  max-width: max-content;
-  min-width: 50%;
-
-  padding: 20px;
-  z-index: 100;
-
-  margin: auto;
-
-  .ticket-details-block-tickets {
-    //border: 2px solid black;
-    display: flex;
-    justify-content: center;
-    //align-items: flex-start;
-    min-width: 40%;
-  }
-
-  .ticket-details {
-    border: 1px solid black;
-    margin: 1em;
-    z-index: 101;
-
-    &__button {
-      z-index: 102;
-      display: block;
-      margin: 5px auto;
-      padding: 5px;
-      border-radius: 5px;
-    }
-  }
-}
