--- src/client/components/pages/booking/flight/TicketDetails.js
@@ -1,23 +1,0 @@
-import React, { Component } from "react";
-
-export class TicketDetails extends Component {
-  render() {
-    return (
-      <div className="ticket-details">
-        Standart Econom Business etc.
-        <br />
-        ticket info
-        <br />
-        price : {this.props.price}
-        <button
-          className="button ticket-details__button"
-          onClick={this.props.onClose}
-        >
-          {this.props.buttonLabel}
-        </button>
-      </div>
-    );
-  }
-}
-
-export default TicketDetails;
